refactoring changes and feedback modifications

+++ Planilla_UCR/DB_Deployment.sql
@@ -34,7 +34,7 @@ CREATE TABLE Project(
 	EmployerEmail varchar(255) NOT NULL,
 	ProjectName varchar(255) NOT NULL,
 	ProjectDescription varchar(255),
-	MaximumAmountForBenefits int, 
+	MaximumAmountForBenefits float, 
 	MaximumBenefitAmount int,
 	PaymentInterval varchar(255),
 	PRIMARY KEY(EmployerEmail, ProjectName),
